auto_scripts/func/requirements.txt
# Python UI & API Automation Framework Dependencies

# Selenium WebDriver
selenium==4.15.2

# Testing Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0

# Configuration Management
PyYAML==6.0.1

# API Testing
requests==2.31.0

# Logging and Reporting
allure-pytest==2.13.2

# WebDriver Manager
webdriver-manager==4.0.1

# Data Handling
python-dotenv==1.0.0

# Utilities
colorama==0.4.6
